+++ pkg/operator/workloadcontroller/apigroup.go
@@ -6,8 +6,24 @@ import (
 
 	"k8s.io/client-go/rest"
 
+	configv1 "github.com/openshift/api/config/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
+
+var apiServiceGroupVersions = []schema.GroupVersion{
+	// these are all the apigroups we manage
+	{Group: "apps.openshift.io", Version: "v1"},
+	{Group: "authorization.openshift.io", Version: "v1"},
+	{Group: "build.openshift.io", Version: "v1"},
+	{Group: "image.openshift.io", Version: "v1"},
+	{Group: "oauth.openshift.io", Version: "v1"},
+	{Group: "project.openshift.io", Version: "v1"},
+	{Group: "quota.openshift.io", Version: "v1"},
+	{Group: "route.openshift.io", Version: "v1"},
+	{Group: "security.openshift.io", Version: "v1"},
+	{Group: "template.openshift.io", Version: "v1"},
+	{Group: "user.openshift.io", Version: "v1"},
+}
 
 func checkForAPIs(restclient rest.Interface, groupVersions ...schema.GroupVersion) []string {
 	missingMessages := []string{}
@@ -22,4 +38,12 @@ func checkForAPIs(restclient rest.Interface, groupVersions ...schema.GroupVersio
 	}
 
 	return missingMessages
+}
+
+func APIServiceReferences() []configv1.ObjectReference {
+	ret := []configv1.ObjectReference{}
+	for _, gv := range apiServiceGroupVersions {
+		ret = append(ret, configv1.ObjectReference{Group: "apiregistration.k8s.io", Resource: "apiservices", Name: gv.Version + "." + gv.Group})
+	}
+	return ret
 }
